基础代码重构 (#122)
* SKR_OS_XXX -> SKR_PLAT_XXX

* FORCEINLINE -> SKR_FORCEINLINE

* SKR_ARCH_XXX

* DECLARE_ZERO -> SKR_DECLARE_ZERO

* fix shadow initializer

* some ThreadSafe code for TypeSystem

* skr_skeleton_resource_t -> skr::anim::SkeletonResource

* use skr::anim namespace

+++ modules/runtime/src/platform/standard/stdio_vfs.cpp
@@ -113,7 +113,7 @@ size_t skr_stdio_fwrite(skr_vfile_t* file, const void* out_buffer, size_t offset
     return -1;
 }
 
-ssize_t skr_stdio_fsize(const skr_vfile_t* file) SKR_NOEXCEPT
+int64_t skr_stdio_fsize(const skr_vfile_t* file) SKR_NOEXCEPT
 {
     if (file)
     {
